chore(aggregate transform): split config and tests
